CasperSD: Rename directory
