+++ src/main.ts
@@ -105,7 +105,7 @@ import * as Scenes_Descent1 from './Descent1_2/Scenes_Descent1.js';
 import * as Scenes_Descent2 from './Descent1_2/Scenes_Descent2.js';
 import * as Scenes_Descent2Vertigo from './Descent1_2/Scenes_Descent2Vertigo.js';
 import * as Scenes_Spyro1 from './Spyro1/scenes.js';
-import * as Scenes_CasperSD from './CasperSD/scenes.js';
+import * as Scenes_CasperSD from './CasperSpiritDimensions/scenes.js';
 
 import { DroppedFileSceneDesc, traverseFileSystemDataTransfer } from './Scenes_FileDrops.js';
